Try another approach: don't run these in parallel
In order to do this, I had to combine the tests into a single module
again.

So why not run in parallel?

If we connect to the same room from multiple parallel tests, the
connection ID increases beyond just 1. This is fine, but there is
currently no convenient way of expressing that.
